feat: 새로고침 시 재연결 로직 hook 분리
